📦 refactor: reorganize models/ into index/model/constants/utils files
- Split each `<domain>-model/index.ts` into a barrel `index.ts` plus a dedicated `model.ts`
- Extract constants (e.g. token expiration times, anonymous features) into `constants.ts`
- Extract helper functions into `utils.ts` where applicable
- Drop redundant `<domain>.model.ts` naming in favor of `model.ts`, since the folder already scopes it
- Update tests that referenced constants through model methods to import the new standalone exports
