+++ src/models/activation-model/constants.ts
@@ -1,0 +1,1 @@
+export const ACTIVATION_TOKEN_EXPIRATION_IN_MILLISECONDS = 60 * 15 * 1000 // 15 minutes
