+++ src/models/authorization-model/constants.ts
@@ -1,0 +1,8 @@
+import { TFeature } from '@/repositories'
+
+export const ANONYMOUS_FEATURES: TFeature[] = [
+  'read:status',
+  'read:activation_token',
+  'create:session',
+  'create:user'
+]
