+++ src/models/session-model/constants.ts
@@ -1,0 +1,1 @@
+export const SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS = 60 * 60 * 24 * 30 * 1000 // 30 Days
