+++ src/models/status-model/model.ts
