+++ src/models/user-model/index.ts
@@ -1,2 +1,2 @@
+export * from './model'
 export * from './schema'
-export * from './user.model'
