+++ src/tests/integration/api/v1/user/get.test.ts
@@ -4,7 +4,10 @@ import { version as uuidVersion } from 'uuid'
 import { API_BASE_URL } from '@/constants'
 import { TUserDto } from '@/dtos'
 import { TErrorResponse } from '@/infra'
-import { sessionModel } from '@/models'
+import {
+  SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS,
+  sessionModel
+} from '@/models'
 import orchestrator from '@/tests/orchestrator'
 import { TApiResponse } from '@/types'
 import { isoStringFieldsToDate } from '@/utils'
@@ -86,7 +89,7 @@ describe('GET /api/v1/user', () => {
       const expectedParsedSetCookie: Cookie = {
         name: 'session_id',
         value: sessionObject.token,
-        maxAge: sessionModel.EXPIRATION_IN_MILLISECONDS / 1000,
+        maxAge: SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS / 1000,
         path: '/',
         sameSite: 'lax',
         httpOnly: true,
@@ -97,7 +100,7 @@ describe('GET /api/v1/user', () => {
 
     test('With halfway-expired session', async () => {
       jest.useFakeTimers({
-        now: new Date(Date.now() - sessionModel.EXPIRATION_IN_MILLISECONDS / 2)
+        now: new Date(Date.now() - SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS / 2)
       })
 
       const createdUser = await orchestrator.createUser({
@@ -149,7 +152,7 @@ describe('GET /api/v1/user', () => {
       const expectedParsedSetCookie: Cookie = {
         name: 'session_id',
         value: sessionObject.token,
-        maxAge: sessionModel.EXPIRATION_IN_MILLISECONDS / 1000,
+        maxAge: SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS / 1000,
         path: '/',
         sameSite: 'lax',
         httpOnly: true,
@@ -198,7 +201,7 @@ describe('GET /api/v1/user', () => {
 
     test('With expired session', async () => {
       jest.useFakeTimers({
-        now: new Date(Date.now() - sessionModel.EXPIRATION_IN_MILLISECONDS)
+        now: new Date(Date.now() - SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS)
       })
 
       const createdUser = await orchestrator.createUser({
